fix : cloudinary error

+++ src/utils/cloudinary.js
@@ -1,5 +1,5 @@
-const { v2 : cloudinary }=require("cloudinary");
-const fs=require("fs")
+const { v2: cloudinary } = require("cloudinary");
+const fs = require("fs");
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -7,18 +7,28 @@ cloudinary.config({
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-async function uploadOnCloudinary(localFilePath,folderPath) {
+async function uploadOnCloudinary(localFilePath, folderPath) {
   if (!localFilePath) return null;
+
   try {
     const uploadResult = await cloudinary.uploader.upload(localFilePath, {
-      folder:folderPath,
+      folder: folderPath,
       resource_type: "auto",
     });
-    fs.unlinkSync(localFilePath);
-    return uploadResult.url;
+
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+
+    return uploadResult.secure_url;
   } catch (error) {
-    fs.unlinkSync(localFilePath);
-    return null
+    console.error("Cloudinary upload failed:", error);
+    
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+
+    return null;
   }
 }
